fix: Keep board motion specs within ESLint budget
The hook tests pushed kanbanCardMotion.spec.ts over the max-lines
limit, so the UI lint job failed the budget check. Move those tests
to their own file so the warning count stays inside the allowed
budget.
